requirements.txt
awscli==1.11.21
awsebcli==3.8.4
beautifulsoup4==4.5.1
blessed==1.9.5
boto==2.43.0
botocore==1.4.78
cement==2.8.2
colorama==0.3.7
defusedxml==0.4.1
Django==1.10.3
django-allauth==0.29.0
django-crontab==0.7.1
django-rest-auth==0.8.2
django-storages==1.5.1
djangorestframework==3.5.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
oauthlib==2.0.1
pathspec==0.3.4
Pillow==3.4.2
psycopg2==2.6.2
pyasn1==0.1.9
pyparsing==2.1.10
python-dateutil==2.6.0
python3-openid==3.0.10
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.7.0
rsa==3.4.2
s3transfer==0.1.9
semantic-version==2.5.0
six==1.10.0
texttable==0.8.7
wcwidth==0.1.7
websocket-client==0.37.0
